Add more functional tests
