+++ src/Components/test/testassets/TestServer/Pages/ConnectionQualityHost.cshtml
@@ -31,6 +31,8 @@
                 logLevel: 1, // LogLevel.Debug
             }).then(function () {
                 window['__aspnetcore__testing__blazor__started__'] = true;
+            }, function () {
+                window['__aspnetcore__testing__blazor__failed__'] = true;
             });
         }
     </script>
